backend/requirements.txt
# Core
Django==5.0.4
djangorestframework==3.15.2
psycopg[binary]==3.2.12

# Celery & Redis
celery==5.3.6
redis==5.0.1

# FastAPI (para orquestração futura)
fastapi==0.111.0
uvicorn==0.30.0

# Utilitários
python-dotenv==1.0.1

# Testes
pytest==8.2.0
pytest-django==4.8.0
